Avoid implicit conversion warning in QtCreator

+++ src/libpentobi_base/BoardConst.cpp
@@ -1088,8 +1088,8 @@ void BoardConst::create_moves(unsigned& moves_created, Piece piece)
     {
         if (log_move_creation)
             LIBBOARDGAME_LOG("Creating moves at ", m_geo.to_string(p));
-        auto x = m_geo.get_x(p);
-        auto y = m_geo.get_y(p);
+        auto x = static_cast<int>(m_geo.get_x(p));
+        auto y = static_cast<int>(m_geo.get_y(p));
         auto point_type = m_geo.get_point_type(p);
         for (size_t i = 0; i < nu_transforms; ++i)
         {
